JAVA_Desktop/Senac_hibernate/src/main/java/br/com/senac/dao/ClienteDaoImpl.java
package br.com.senac.dao;

import br.com.senac.entidade.Cliente;
import java.io.Serializable;
import java.util.List;
import org.hibernate.HibernateException;
import org.hibernate.Session;
import org.hibernate.query.Query;

public class ClienteDaoImpl extends BaseDaoImpl<Cliente, Long> implements ClienteDao, Serializable {

    @Override
    public Cliente pesquisarPorId(Long id, Session sessao) throws HibernateException {
        return sessao.find(Cliente.class, id);
    }

    @Override
    public List<Cliente> pesquisarPorNome(String nome, Session sessao) throws HibernateException {
        Query<Cliente> consulta = sessao.createQuery("FROM Cliente u WHERE u.nome LIKE :vNome");    //u.nome = getnome
        consulta.setParameter("vNome", "%" + nome + "%");
        return consulta.getResultList();
    }

}
